修改整数溢出问题
# message auto-generated for no-merge-commit merge:
merge master_fix_file into master

修改整数溢出问题

Created-by: insightDeveloper

Match-id-47552c6c9f8b976324683f58e27007e5485bf7f8

+++ server/msinsight/src/utils/NumberUtil.h
@@ -42,7 +42,7 @@ public:
 
     static inline unsigned long long TryParseUnsignedLongLong(const std::string &longLongStr)
     {
-        int ret;
+        unsigned long long ret;
         try {
             ret = std::stoull(longLongStr);
         } catch (std::invalid_argument &) {
